refactor internals to force some required keys
