+++ src/runtime/server/apollo-server/express.ts
@@ -1,0 +1,8 @@
+import { ApolloServer, ApolloServerExpressConfig } from 'apollo-server-express'
+import { ApolloConfig } from './types'
+
+export class ApolloServerExpress extends ApolloServer {
+  constructor(config: ApolloConfig<ApolloServerExpressConfig>) {
+    super(config)
+  }
+}
